implement glfw into our engine. Peace glut you will not be missed.

+++ src/Engine/stateManager.c
@@ -87,10 +87,10 @@ int StateManager_keyUp(StateManager *stateManager, InputType inputType) {
     return 1;
 }
 
-int StateManager_mouseMove(StateManager *stateManager, float x, float y) {
+int StateManager_mouseMove(StateManager *stateManager, double x, double y) {
     State *state = StateManager_top(stateManager);
     if (state->mouseMovement != NULL) {
-        return state->mouseMovement(x, y);
+        return state->mouseMovement((float)x, (float)y);
     }
     return 1;
 }
